Incorporating "gr" functions into sp module to make new spgr module.
gr_find() writes to an output file all the OGs of a query and each of the other eukaryote groups that also have that OG.
Frequencies of group occurences can then be computed with gr_count().
In other words, these functions compare the total shared no. of OGs (extent of shared genome) of each group with the query group.